Move extension health check types
