+++ hack/api-reference/config.json
@@ -20,8 +20,8 @@
             "docsURLTemplate": "https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.15/#{{lower .TypeIdentifier}}-{{arrIndex .PackageSegments -1}}-{{arrIndex .PackageSegments -2}}"
         },
         {
-            "typeMatchPrefix": "github.com/gardener/gardener/extensions/pkg/controller/healthcheck/config",
-            "docsURLTemplate": "https://github.com/gardener/gardener/extensions/pkg/controller/healthcheck/config"
+            "typeMatchPrefix": "github.com/gardener/gardener/extensions/pkg/apis/config",
+            "docsURLTemplate": "https://github.com/gardener/gardener/extensions/pkg/apis/config"
         }
     ],
     "typeDisplayNamePrefixOverrides": {
