+++ hack/api-reference/config.md
@@ -74,9 +74,7 @@ ETCD
 <td>
 <code>healthCheckConfig</code></br>
 <em>
-<a href="https://github.com/gardener/gardener/extensions/pkg/controller/healthcheck/config">
-github.com/gardener/gardener/extensions/pkg/controller/healthcheck/config/v1alpha1.HealthCheckConfig
-</a>
+invalid type
 </em>
 </td>
 <td>
